+++ src/main/kotlin/dev/nyon/autodrop/config/Config.kt
@@ -49,7 +49,7 @@ data class Archive(
 )
 
 /**
- * Represents and entry of an archive for identification of items that are valid to be dropped.
+ * Represents an entry of an archive for identification of items that are valid to be dropped.
  *
  * @param type is the type of the item and serialized as a [dev.nyon.autodrop.extensions.ResourceLocation], e.g. minecraft:stone.
  * @param predicate is the item data to be filtered for. The format should match the syntax of the item predicate.