+++ src/main/kotlin/dev/nyon/autodrop/extensions/Component.kt
@@ -1,5 +1,7 @@
 package dev.nyon.autodrop.extensions
 
+import net.minecraft.core.component.DataComponents
+import net.minecraft.network.chat.CommonComponents
 import net.minecraft.network.chat.Component
 import net.minecraft.network.chat.MutableComponent
 import net.minecraft.world.item.Item
@@ -10,9 +12,5 @@ fun screenComponent(key: String, vararg objects: Any): MutableComponent {
 
 val Item.narration: Component
     get() {
-        //? if >=1.21.2
-        return this.name
-
-        //? if <1.21.2
-        /*return Component.literal(this.description.toString())*/
+        return components().getOrDefault(DataComponents.ITEM_NAME, CommonComponents.EMPTY)
     }