+++ changelog.md
@@ -1,3 +1,4 @@
-- config screen:
-  - Add a hint for enchantment predicates in ModifyScreen
-  - allow saving an identifier with an invalid predicate [**#31**](https://github.com/btwonion/autodrop/issues/31)
+- update to 26.1
+    - since the prereleases of this version, it’s been no longer possible to access items when no world is loaded
+        <br>-> Due to this, the config screen will only work when you are in a world
+- dropped all support for versions <26.1